Factor out NoteRevisionMessage
